moved some of the key functions out of the 'block_operations' in preparation for deletion
